add auth for authentication of user in methods store, edit, update and destroy

+++ app/Http/Controllers/Admin/PhotoController.php
@@ -18,7 +18,7 @@ class PhotoController extends Controller
     public function index()
     {
        //dd(Photo::orderByDesc('id')->paginate());
-       return view('admin.photos.index', ['photos' => Photo::orderByDesc('id')->paginate()]);
+       return view('admin.photos.index', ['photos' => Photo::where('user_id', auth()->id())->orderByDesc('id')->paginate()]); // auth() acceso all'utente attualmente registrato
     }
 
     /**
@@ -42,6 +42,10 @@ class PhotoController extends Controller
         $image_path = Storage::put('uploads', $request->upload_image);
         $val_data['upload_image'] = $image_path;
         // dd($val_data);
+
+        // authenticazione dell utente
+        $val_data['user_id'] = auth()->id(); // validation, il campo user_id é uguale all id dell utente registrato
+
         Photo::create($val_data);
         return to_route('admin.photos.index')->with('message', 'Foto aggiunta con successo !');
     }
@@ -60,6 +64,11 @@ class PhotoController extends Controller
      */
     public function edit(Photo $photo)
     {
+        // authenticazione dell utente
+        if(auth()->id() != $photo->user_id){ // Se l'id del utente registrato non é uguale all user_id di quell post
+            abort(403, 'Questo account non coincide con il tuo'); // allora errore
+        }
+        
         $categories = Category::all();
         return view('admin.photos.edit', compact('photo', 'categories'));
     }
@@ -69,20 +78,25 @@ class PhotoController extends Controller
      */
     public function update(UpdatePhotoRequest $request, Photo $photo)
     {
-       // dd($request->all());
-       $val_data = $request->validated();
-       $val_data['slug'] = Str::slug('title', '-');
+        // authenticazione dell utente
+        if(auth()->id() != $photo->user_id){ // Se l'id del utente registrato non é uguale all user_id di quell post
+            abort(403, 'Questo account non coincide con il tuo'); // allora errore
+        }  
 
-       if ($request->has('upload_image')) { // if 1:  se cé una upload_image nell mio aggiornamento($request)
-        
-            if ($photo->upload_image) { // if 2: controllare se gia cera una upload_image 
-                
-                Storage::delete($photo->upload_image); //if 2: allora cancellare
-            } 
-        
-            $image_path = Storage::put('uploads', $request->upload_image); // if 1 :  allora recuperare l'immagine e salvarla nel DB
-            $val_data['upload_image'] = $image_path;   
-    }
+        // dd($request->all());
+        $val_data = $request->validated();
+        $val_data['slug'] = Str::slug('title', '-');
+
+        if ($request->has('upload_image')) { // if 1:  se cé una upload_image nell mio aggiornamento($request)
+            
+                if ($photo->upload_image) { // if 2: controllare se gia cera una upload_image 
+                    
+                    Storage::delete($photo->upload_image); //if 2: allora cancellare
+                } 
+            
+                $image_path = Storage::put('uploads', $request->upload_image); // if 1 :  allora recuperare l'immagine e salvarla nel DB
+                $val_data['upload_image'] = $image_path;   
+        }
 
        $photo->update($val_data);
 
@@ -94,6 +108,11 @@ class PhotoController extends Controller
      */
     public function destroy(Photo $photo)
     {
+        // authenticazione dell utente
+        if(auth()->id() != $photo->user_id){
+            abort(403, 'Questo account non coincide con il tuo'); 
+        }
+
         if ($photo->upload_image) { // questo if lo abbiamo portato da update
             // if so, delete it
             Storage::delete($photo->upload_image);
